refactor: Address code quality issues from PR review
Addresses Copilot PR review feedback for code quality improvements:

1. **Eliminate Code Duplication**
   - Created shared `lib/environment-utils.js` module
   - Extracted `getEnvironment()` and `getBaseUrl()` functions
   - Updated `build-config.js` and `set-base-url.js` to use shared module
   - DRY principle: Single source of truth for environment logic

2. **Improve Regex Robustness (set-base-url.js)**
   - Updated base href regex to handle attributes in any order
   - Pattern: `/<base\s+[^>]*href="[^"]*"[^>]*>/i`
   - Now handles cases like `<base id="x" href="...">`

3. **Fix data-environment Attribute Logic (set-base-url.js)**
   - More specific check for existing attribute on <body> tag
   - Updates existing attribute instead of duplicating
   - Prevents false positives from string matches elsewhere in HTML

4. **Clean Build Artifacts**
   - index.html reset to clean state (no build-time modifications)
   - Build artifacts (base href, data-environment, cache versions)
   - Now generated only at build-time, not committed to git

Testing:
- ✓ npm run build:production - works correctly
- ✓ npm run build:staging - works correctly
- ✓ npm run build:testing - works correctly
- ✓ Shared module imported successfully
- ✓ Improved regex handles edge cases

Note: Firebase security concerns (credentials in git, environment isolation)
are intentionally deferred to Phase 3 (Issue #110) as originally planned.

Related: #111 (PR #112)
